requirements.txt
absl-py==0.7.1
addict==2.4.0
appdirs==1.4.4
astunparse==1.6.3
Bottleneck==1.2.1
certifi==2019.9.11
chardet==3.0.4
cityscapesScripts==2.2.1
click==7.1.2
colorama==0.4.6
coloredlogs==15.0.1
cycler==0.10.0
Cython==0.29.16 
cython-bbox==0.1.3  
filelock==3.12.0
gast==0.3.3
git-lfs==1.6
gitdb==4.0.10
GitPython==3.1.31
google-pasta==0.1.8
grpcio==1.27.2
huggingface-hub==0.13.4
humanfriendly==10.0
idna==2.8
imagecorruptions==1.1.2
imageio==2.27.0
importlib-metadata==6.6.0
joblib==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lap==0.4.0
Markdown==3.4.3
markdown-it-py==2.2.0
matplotlib==3.2.2
mdurl==0.1.2
mmcv-full==1.3.18
mmdet==2.18.0
mmengine==0.7.2
model-index====0.1.11
mpmath==1.1.0
networkx==2.6.3
numexpr==2.7.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
openmim==0.1.1
opt-einsum==3.1.0
ordered-set====4.1.0
packaging==23.1
pandas==1.0.5
panopticapi====0.1
Pillow==9.5.0
pip==23.1.1
ply==3.11
protobuf==3.11.2
pycocotools====2.0
Pygments==2.15.1
pyparsing==2.4.2
pyquaternion==0.9.9
python-dateutil==2.8.0
python-git==2018.2.1
pytz==2019.3
PyWavelets==1.3.0
PyYAML==5.3.1
requests==2.22.0
rich==13.3.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.5.0
seaborn==0.12.2
semver==2.8.1
Send2Trash==1.8.0
setuptools==67.7.1
shapely==2.0.1
six==1.14.0
smmap==5.0.0
sympy==1.4
tabulate==0.9.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.13
tomli==2.0.1
torch==1.10.2
torchvision====0.11.3
tqdm==4.65.0
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.25.6
virtualenv==16.7.6
wheel==0.40.0
wrapt==1.11.2
xarray==0.14.0
yapf==0.33.0
zipp==3.15.0